requirements.txt
absl-py==0.11.0
alembic==1.4.3
aniso8601==8.1.0
astroid @ file:///tmp/build/80754af9/astroid_1592495881661/work
astunparse==1.6.3
beautifulsoup4==4.9.3
bidict==0.21.2
cached-property==1.5.2
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
eventlet==0.30.0
filelock==3.0.12
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Session==0.3.2
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-self-attention==0.47.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
matplotlib==3.4.2
mccabe==0.6.1
networkx==2.5.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint @ file:///tmp/build/80754af9/pylint_1598624038450/work
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2020.1
pyvi==0.1
PyWavelets==1.1.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
sklearn-crfsuite==0.3.6
soupsieve==2.0.1
SQLAlchemy==1.3.20
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.12.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tf2crf==0.1.31
threadpoolctl==2.1.0
tifffile==2021.6.6
tokenizers==0.10.1
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tqdm==4.51.0
transformers==4.4.1
typed-ast==1.4.1
typeguard==2.11.1
urllib3==1.26.2
vncorenlp==1.0.3
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.3
zipp==3.4.0
